requirements.txt
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
Deprecated==1.2.13
importlib-metadata==4.11.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
redis==4.3.2
rsa==4.8
Telethon==1.24.0
tomli==2.0.1
typing_extensions==4.2.0
wrapt==1.14.1
zipp==3.8.0
